Remove obsolete crutest perf test
Remove the perf test from crutest.  There are other better tests now and
this one is just a distraction.

Removed the tools/test_perf.sh script and places that referenced it.

+++ upstairs/src/dummy_downstairs_tests.rs
@@ -1596,7 +1596,7 @@ async fn test_byte_fault_condition_offline() {
         h.await.unwrap();
 
         let ds = harness.guest.downstairs_state().await.unwrap();
-        if (i as usize + 1) * WRITE_SIZE < IO_OUTSTANDING_MAX_BYTES as usize {
+        if (i + 1) * WRITE_SIZE < IO_OUTSTANDING_MAX_BYTES as usize {
             assert_eq!(ds[ClientId::new(0)], DsState::Offline);
             assert_eq!(ds[ClientId::new(1)], DsState::Active);
             assert_eq!(ds[ClientId::new(2)], DsState::Active);
